bot/requirements.txt
aiogram
requests
python-dotenv
aiohttp
deepl
google-generativeai
httpx
matplotlib